Add position property to avfilter (#345).
"frame" (default) is the position of the frame within the composition.
This corresponds to the consumer's render position.
"filter" is relative to the the filter's in point.
"producer" is relative to the producer's in point (producer here could
be a playlist or multitrack).
"source" is the position from the original producer that generated this
frame (i.e. absolute).