Remove RGB conversion in vid.stab.
If a filter does not directly support an image format, it should not try
to convert itself. Rather, it should let the assigned image converter do
it. I started retesting vid.stab native support for RGB and RGBA only to
have it crash. Upon debugging and reviewing the vid.stab source, I see
it does not support the configure interpolation method for packed images.
Rather, it always uses bilinear for packed.