Add mlt_luma_map.
Several services across a few modules use this functionality. So, it
made sense to centralize it. However, the main motivation for this was
to remove the need to generate luma files in the luma module and let
these images be generated just-in-time. This removes a large amount of
data from install and makes them more flexible to image resolution and
aspect ratio.