fix serializing animated property with new length
When setting a keyframe value using like mlt_property_anim_set_... after
the first time, the length is ignored. Since the animation length was
not updated, mlt_animation_serialize_cut_tf() ends sooner than expected,
and the animation is not completely serialized.

See https://forum.shotcut.org/t/keyframes-added-to-extended-parts-of-
clips-are-made-invalid/13305