fix video transition with video clip on audio track
When hunting for a valid A frame, it was only checking for valid audio
or video per the transition type. It was not checking whether the audio
or video on the frame was hidden. This adds a check against the "hide"
property on the frame applied from the track's "hide" property.

See https://forum.shotcut.org/t/black-background-with-image-overlay/
20343/