fix #590 bad interpolation when scaling with affine rect
We now request upstream to supply a scaled image unless not using fill or
distort modes and the image is smaller than the rect (unscaled use
case). The overall strategy is not scale with a dedicated image scaler
and use affine with sub-pixel precision via interpolation for positioning,
rotation, and shear.