improve slices for frei0r
This fixes using threads=0 for automatic slice count and now works with
images that are not a multiple of the specified slice count by
automatically reducing the slice count as needed. You still must opt
into using slices by setting the threads property because there are
many plugins with which it does not work.