drop the special fs: protocol
This was previously required to be able to supply a query string of
additional open parameters on a file name. Now, an escaped question
mark is considered a query string delimiter on file names: \?.
It must be esacped to accept a question in the file name.
If the resource URL begins with a protocol other than "file:"
then the query string delimiter continues to be an unescaped
question mark.