Do not load profile until needed in chain creation
When creating the native producer profile, the avformat produer will
be initialized by reading images. This needs to wait until the last
opportunity in case the user/application has a need to set more
properties on it before it is initialized.

This also reduces memory use for chains that do not have links.

Fixes:
https://forum.shotcut.org/t/changing-color-range-doesnt-affect-source-player/27788