add mlt_frame_get_alpha_size()
This fixes a crash when cloning a frame with rgba that has an alpha
channel property with unknown or incorrect byte size.
See https://forum.shotcut.org/t/square-video-clips-cause-a-crash/33693

Basically, a frame that has an rgba image should not have a distinct
alpha channel because it is likely to not be synchronized or at least
redundant. In lieu of chasing down all the ways image is set and
clearing the alpha property when needed, this change prevents using a
distinct alpha channel when there is rgba.