add filter-relative timing to opencv.tracker
opencv.tracker's timing is different than other dual pass filters such as
vidstab and loudness, which are relative to the filter. Rather, it is
relative to the original producer. When using a chain in MLT XML as the
filter's producer, tracker's first pass analysis mode is broken and never
completes because the producer length is a maximal value. Instead of
trying to somehow fix that elsewhere (in a chain the duration of the
producer and attached filters are fluid by design and the original
producer duration is less relevant), this change introduces a timing
mode that works the same as the other filters when used with a chain.