Add link_avdeinterlace and link_deinterlace (#878)
* Add link_avdeinterlace and link_deinterlace

* Resolve review comments

Rename deinterlace_method to deinterlacer

* Fix deinterlace cpu acceleration

CPU acceleration was not being enabled for x86_64.

* Multiple fixes for deinterlace filter

* Use correct text string for yadif-nospatial
* Provide next frame for weave and greedy

* Reorder deinterlacer priority based on performance testing

* Use mlt_producer_probe() to detect progressive source in get_frame()

This allows deinterlacers to avoid calling get image at all if the
source is progressive.

* Rename MLT_SWS_FLAGS to MLT_AVFILTER_SWS_FLAGS

This helps to distinguish from mlt_get_sws_flags() which is preferred
when the source and destination sws parameters are known. But does
not work when giving avfilter hints about how to create automatic
scaler filters.

* Ignore qt6 autogenerated files

* Add option to list links and link metadata

* Fix metadata for multiple services is not registered or installed correctly

* Do not normalize a link until after the source has been added

This allows the chain to remove the normalizing filters from the
source producer if it wants to.

* Remove size "test" parameter from swscale and avcolor_space

The swscale resolution test is no longer needed as of FFmpeg v0.7.
This was already removed from producer_loader in 2014:
https://github.com/mltframework/mlt/commit/2fc495f6596fc75d5ec0245dbe0fdf63e4a1dee7

* Add chain normalizers

The new link_filter functions make an easy way to wrap existing
normalizer filters into links. With this change, chains to not need
to have normalizer filters on the chain or the encapsulated source
producer if it has had normalizer links applied.

* Remove normalizing filters from deinterlace links

The normalize links will now do the normalizing.

* Defer relinking chain until a frame is required.

This reduces relinking when multiple changes are made (for example,
adding normalizing links).

* Fix progressive source detection

* Put new function definitions in the new version block

* Ensure normalizers are only added to chains once when loaded from xml

* Make deinterlace detection more robust

Test for progressive source with every frame
Test for progressive consumer with every image
This is a little less efficient, but it is more robust in responding
to changes in the interlace source or display

* Request native audio format in swresample

This avoides a format conversion since swresample can convert
any format.

* Fix crash on opening a project with audio only clips

Do not lock the mutextes because they may not exist yet.

* The convert filter must be first so that other normalizers can use it.

* Fix audio tick at clip splits

Correctly calculate the original frame number when using future frames
to load the sample rate converter.

* Fix timeremap does not work with GPU effects

* Fix crash when using timeremap with movit