fix crash in count producer via xml producer
regression in 1d63bd8

The xml consumer was serializing the mlt_producer default "resource"
property value of "<producer>", which the xml producer would prepend
with a root path. The loader producer would append the resource value to
the service name delimited by a colon. That was then split and sent to
the count producer's init function as an invalid argument to use for the
factory producer.