transition_affine: Fix of memory overconsumption.
When I montaged my cool MLG-montage using Shotcut, I noticed that: when I zoom using "affineSizePosition" filter, memory consumption increases dramatically and it eats all of my RAM and swap file! And it's also very very slow!
But c'mon, that shouldn't eat so much. And I thought that this filter just allocates a huge buffer with the size of the zoomed frame instead of displaying buffer of normal size with applied scale. And there was exactly that thing.
Now we can zoom infinitely! And now it's also fast!
And thank you for the art freedom!