Fix mask start/apply corruption by intermediate filters
When using mask_start and mask_apply, filters between them can
corrupt the mask. This patch stores the mask on the clone frame
so that intermediate filters can not corrupt it. The cloned frame
mask is inverted and the coposite is reversed to achieve the same
effect.

Originally reported here:
https://forum.shotcut.org/t/beta-version-23-04-now-available-to-test/38269/125