Fix a possible crash in avformat probe (#944)
* Fix a possible crash in avformat probe

Probe was accessing the cached avformat context from a store pointer
(self->child) rather than through the cache. This can result in
accesssing released memory if the cache deleted the context.

The context pointer should not be saved at all and should only be
accessed through the cache system.

This change includes a rewrite of the video index change detection
so that probe() will not have to access the avformat context at all.

This also adds an optimization that if metadata is loaded from XML,
and nothing else has changed, then the probe can be skipped.

* Use meta.media.nb_streams instead of meta.media.stream.0.codec

Stream 0 may not exist. And nb_streams should always be present