Fix #948 Crash when avformat producer properties change
The property_changed event passes a pointer to the producer_avformat
object that listens to the event. But that producer_avformat instance
could have been deleted by the image crash. So the event callback
tries to access previously released memory.

This change disconnects the event callback when the producer_avformat
object is destroyed.