fix explicit audio_index and video_index not used
The problem occurs when the mlt_service is `avformat` instead of
`avformat-novalidate` and `audio_index` or `video_index` is supplied
that is different than default but not `astream`/`vstream` properties
(legacy project) properties. In that case, the explicit `audio_index` or
`video_index` is ignored because `astream` and `vstream` are set to some
default values. In that case, the project or command line fails to play
the correct stream. TO reproduce `melt some-media-file audio_index=2`,
for example (assuming there is a second audio track at index 2).

regression in 52c5bde