fix movit transition with non-movit filter on input
The chain fingerprint code did not handle the situation where one of its
inputs could change. In this case, it was a transition whose inputs were
not from the same chain: a non-movit filter interrupts it.
`filter_movit_convert.cpp:set_movit_parameters()` would fail to locate
the input, not set the pixel data, and return making rendering non-
deterministic.

https://forum.shotcut.org/t/text-simple-bug-with-gpu-efx-on/43892