Fix audio does not play for audio only file
This specific file was a WMA file with DTS coded audio. The first
packet of audio does not have a PTS of zero. Since there is no
video in the file, MLT tries to seek to position 0 for the first
audio frame and attempts to sync audio frames as if they started
at 0. This patch finds a first_pts value if there is no video in
the file to look for.