Fix mask_apply with custom transition
When using transition=frei0r.cairoblend or qtblend and multiple
mask_apply with filter=frei0r.alphaspot, for example, to make compound
shapes, the compositing does not work correctly. It would have a black
background instead of the original frame. I found it has something to do
with cloning a frame that was already cloned. Also, it is wasteful to
clone for each mask_start; only the first one is needed.