Use CMake's find_package for fftw3 (#1114)
If FFTW3 itself was built using CMake it comes with native support for find_package, otherwise not. Hence we need to add a fallback to pkg-config which is available for all build systems, but is not cross-plattform compatible. Namely on MSVC it tries to link to m (which is not available on MSVC)