Add mlt types for colorspace, trc, and primaries (#1155)
* Add mlt type for color_trc

The enum value matches the FFMpeg values for backwards compatibility

* Put mlt_colorspace type to use

Change the enum values to match the hard coded values in the code.
Add string conversions so colorspace can be specified a number or
a name. Convert the code to use the mlt_colorspace type.

* Add mlt type for color_primaries

The enum value matches previous values that were hard coded for
backwards compatibility. For new values, the values match FFMpeg
values

* Set mlt_color_primaries values to match FFMpeg

* Fix tests