Don't change colorspace/range on mlt_image_none
This image format is supposed to instruct the producer to provide as
native or closest to source as possible. That can be useful by API but
also Movit uses this because it tries to handle things as much as
possible. This improves the color accuracy of SDR within a BT.2020/HLG
project on Movit.