Add sws_colortransform filter
This reduces one memory copy over the avfilter.scale version. On my
system, it is faster than avfilter.zscale on mlt_image_format=rgba64 but
not mlt_image_format=rgba. Prioritizing for 8-bit since that is the
majority. Still, this is a better fallback when zscale is not included
in the FFmpeg build.