Add vcpkg related files and MSVC CI job (#1189)
* Add vcpkg files

While vcpkg supports Linux, our main focus is on Windows (MSVC) for now. In this first step a minimal set of dependencies is added to be able to do a build of the framework with most modules disabled.

* Add CMakePresets.json

This makes it easier to run CMake with a fixed set of many options. It is also a preperations step for a vcpkg based CI job.

* Add CI job for MSVC powered by vcpkg

This is an inital draft that rebuilds all dependencies all the time. For use in production caching should be added (either via GitHub Action Cache or via GitHub Packages).